Split plugin adapter mutex in two and use them to lock only on data access
